+++ src/jarvis/jarvis-config.ts
@@ -4,7 +4,7 @@ const JARVIS_CONFIG: {
     CYCLE_TIME: 60 * 1000, //ms
 };
 
-const NETWORK_CONFIG: {
+export const NETWORK_CONFIG: {
     BLACK_LIST: string[],
 } = {
     BLACK_LIST: ['home', 'darkweb', 'CSEC', 'The-Cave', 'run4theh111z', 'I.I.I.I', 'avmnite-02h', '.', 'w0r1d_d43m0n'],